Create and populate HW and Comb to LLVM pass files
