+++ lib/Conversion/LLHDToLLVM/CombToLLVM.cpp
@@ -1,0 +1,216 @@
+//===- CombToLLVM.cpp - Comb to LLVM Conversion Pass ----------------------===//
+//
+// Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
+// See https://llvm.org/LICENSE.txt for license information.
+// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+//
+//===----------------------------------------------------------------------===//
+//
+// This is the main Comb to LLVM Conversion Pass Implementation.
+//
+//===----------------------------------------------------------------------===//
+
+#include "circt/Conversion/HWToLLVM.h"
+#include "../PassDetail.h"
+#include "circt/Dialect/Comb/CombOps.h"
+#include "circt/Dialect/HW/HWOps.h"
+#include "circt/Dialect/LLHD/IR/LLHDDialect.h"
+#include "circt/Dialect/LLHD/IR/LLHDOps.h"
+#include "circt/Support/LLVM.h"
+#include "mlir/Conversion/ControlFlowToLLVM/ControlFlowToLLVM.h"
+#include "mlir/Conversion/FuncToLLVM/ConvertFuncToLLVM.h"
+#include "mlir/Conversion/FuncToLLVM/ConvertFuncToLLVMPass.h"
+#include "mlir/Conversion/LLVMCommon/ConversionTarget.h"
+#include "mlir/Conversion/LLVMCommon/Pattern.h"
+#include "mlir/Conversion/ReconcileUnrealizedCasts/ReconcileUnrealizedCasts.h"
+#include "mlir/Dialect/Func/IR/FuncOps.h"
+#include "mlir/Dialect/LLVMIR/LLVMDialect.h"
+#include "mlir/IR/BlockAndValueMapping.h"
+#include "mlir/IR/BuiltinOps.h"
+#include "mlir/IR/BuiltinTypes.h"
+#include "mlir/Pass/Pass.h"
+#include "mlir/Transforms/DialectConversion.h"
+#include "llvm/ADT/TypeSwitch.h"
+
+using namespace mlir;
+using namespace circt;
+using namespace circt::llhd;
+
+//===----------------------------------------------------------------------===//
+// Extraction operation conversions
+//===----------------------------------------------------------------------===//
+
+namespace {
+/// Convert a comb::ExtractOp to LLVM dialect.
+struct CombExtractOpConversion : public ConvertToLLVMPattern {
+  explicit CombExtractOpConversion(MLIRContext *ctx,
+                                   LLVMTypeConverter &typeConverter)
+      : ConvertToLLVMPattern(comb::ExtractOp::getOperationName(), ctx,
+                             typeConverter) {}
+
+  LogicalResult
+  matchAndRewrite(Operation *op, ArrayRef<Value> operands,
+                  ConversionPatternRewriter &rewriter) const override {
+    auto extractOp = cast<comb::ExtractOp>(op);
+    mlir::Value valueToTrunc = extractOp.getInput();
+    mlir::Type type = extractOp.getInput().getType();
+
+    if (extractOp.getLowBit() != 0) {
+      mlir::Value amt = rewriter.create<LLVM::ConstantOp>(
+          op->getLoc(), type, extractOp.getLowBitAttr());
+      valueToTrunc = rewriter.create<LLVM::LShrOp>(op->getLoc(), type,
+                                                   extractOp.getInput(), amt);
+    }
+
+    rewriter.replaceOpWithNewOp<LLVM::TruncOp>(
+        op, extractOp.getResult().getType(), valueToTrunc);
+    return success();
+  }
+};
+} // namespace
+
+//===----------------------------------------------------------------------===//
+// Concat operations conversion
+//===----------------------------------------------------------------------===//
+
+namespace {
+/// Convert a comb::ConcatOp to the LLVM dialect.
+struct CombConcatOpConversion : public ConvertToLLVMPattern {
+  explicit CombConcatOpConversion(MLIRContext *ctx,
+                                  LLVMTypeConverter &typeConverter)
+      : ConvertToLLVMPattern(comb::ConcatOp::getOperationName(), ctx,
+                             typeConverter) {}
+
+  LogicalResult
+  matchAndRewrite(Operation *op, ArrayRef<Value> operands,
+                  ConversionPatternRewriter &rewriter) const override {
+    auto concatOp = cast<comb::ConcatOp>(op);
+    auto numOperands = concatOp->getNumOperands();
+    mlir::Type type = concatOp.getResult().getType();
+
+    unsigned nextInsertion = type.getIntOrFloatBitWidth();
+    auto aggregate = rewriter
+                         .create<LLVM::ConstantOp>(op->getLoc(), type,
+                                                   IntegerAttr::get(type, 0))
+                         .getRes();
+
+    for (unsigned i = 0; i < numOperands; i++) {
+      nextInsertion -=
+          concatOp->getOperand(i).getType().getIntOrFloatBitWidth();
+
+      auto nextInsValue = rewriter.create<LLVM::ConstantOp>(
+          op->getLoc(), type, IntegerAttr::get(type, nextInsertion));
+      auto extended = rewriter.create<LLVM::ZExtOp>(op->getLoc(), type,
+                                                    concatOp->getOperand(i));
+      auto shifted = rewriter.create<LLVM::ShlOp>(op->getLoc(), type, extended,
+                                                  nextInsValue);
+      aggregate =
+          rewriter.create<LLVM::OrOp>(op->getLoc(), type, aggregate, shifted)
+              .getRes();
+    }
+
+    rewriter.replaceOp(op, aggregate);
+    return success();
+  }
+};
+} // namespace
+
+//===----------------------------------------------------------------------===//
+// Arithmetic conversions
+//===----------------------------------------------------------------------===//
+
+namespace {
+
+using CombAddOpConversion = VariadicOpConversion<comb::AddOp, LLVM::AddOp>;
+using CombMulOpConversion = VariadicOpConversion<comb::MulOp, LLVM::MulOp>;
+using CombSubOpConversion =
+    OneToOneConvertToLLVMPattern<comb::SubOp, LLVM::SubOp>;
+
+using CombDivUOpConversion =
+    OneToOneConvertToLLVMPattern<comb::DivUOp, LLVM::UDivOp>;
+using CombDivSOpConversion =
+    OneToOneConvertToLLVMPattern<comb::DivSOp, LLVM::SDivOp>;
+
+using CombModUOpConversion =
+    OneToOneConvertToLLVMPattern<comb::ModUOp, LLVM::URemOp>;
+using CombModSOpConversion =
+    OneToOneConvertToLLVMPattern<comb::ModSOp, LLVM::SRemOp>;
+
+using CombICmpOpConversion =
+    OneToOneConvertToLLVMPattern<comb::ICmpOp, LLVM::ICmpOp>;
+
+// comb.mux supports any type thus this conversion relies on the type converter
+// to be able to convert the type of the operands and result to an LLVM_Type
+using CombMuxOpConversion =
+    OneToOneConvertToLLVMPattern<comb::MuxOp, LLVM::SelectOp>;
+
+} // namespace
+
+namespace {
+/// Convert a comb::ParityOp to the LLVM dialect.
+struct CombParityOpConversion : public ConvertToLLVMPattern {
+  explicit CombParityOpConversion(MLIRContext *ctx,
+                                  LLVMTypeConverter &typeConverter)
+      : ConvertToLLVMPattern(comb::ParityOp::getOperationName(), ctx,
+                             typeConverter) {}
+
+  LogicalResult
+  matchAndRewrite(Operation *op, ArrayRef<Value> operands,
+                  ConversionPatternRewriter &rewriter) const override {
+    auto parityOp = cast<comb::ParityOp>(op);
+
+    auto popCount =
+        rewriter.create<LLVM::CtPopOp>(op->getLoc(), parityOp.getInput());
+    rewriter.replaceOpWithNewOp<LLVM::TruncOp>(
+        op, IntegerType::get(rewriter.getContext(), 1), popCount);
+
+    return success();
+  }
+};
+
+} // namespace
+
+//===----------------------------------------------------------------------===//
+// Bitwise conversions
+//===----------------------------------------------------------------------===//
+
+namespace {
+template <typename SourceOp, typename TargetOp>
+class VariadicOpConversion : public ConvertOpToLLVMPattern<SourceOp> {
+public:
+  using OpAdaptor = typename SourceOp::Adaptor;
+  using ConvertOpToLLVMPattern<SourceOp>::ConvertOpToLLVMPattern;
+  using Super = VariadicOpConversion<SourceOp, TargetOp>;
+
+  LogicalResult
+  matchAndRewrite(SourceOp op, OpAdaptor adaptor,
+                  ConversionPatternRewriter &rewriter) const override {
+
+    size_t numOperands = op.getOperands().size();
+    // All operands have the same type.
+    Type type = op.getOperandTypes().front();
+    auto replacement = op.getOperand(0);
+
+    for (unsigned i = 1; i < numOperands; i++) {
+      replacement = rewriter.create<TargetOp>(op.getLoc(), type, replacement,
+                                              op.getOperand(i));
+    }
+
+    rewriter.replaceOp(op, replacement);
+
+    return success();
+  }
+};
+
+using AndOpConversion = VariadicOpConversion<comb::AndOp, LLVM::AndOp>;
+using OrOpConversion = VariadicOpConversion<comb::OrOp, LLVM::OrOp>;
+using XorOpConversion = VariadicOpConversion<comb::XorOp, LLVM::XOrOp>;
+
+using CombShlOpConversion =
+    OneToOneConvertToLLVMPattern<comb::ShlOp, LLVM::ShlOp>;
+using CombShrUOpConversion =
+    OneToOneConvertToLLVMPattern<comb::ShrUOp, LLVM::LShrOp>;
+using CombShrSOpConversion =
+    OneToOneConvertToLLVMPattern<comb::ShrSOp, LLVM::AShrOp>;
+
+} // namespace
